# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CyberVerse AI - Hacker vs Defender Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /score:
    post:
      operationId: submitScore
      tags: [scores]
      summary: Submit score and gain XP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreBody"
      responses:
        "200":
          description: Score submitted and XP updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [scores]
      summary: Get aggregated dashboard stats for current user
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /skills:
    get:
      operationId: getSkills
      tags: [skills]
      summary: Get all skills and user unlock status
      responses:
        "200":
          description: Skills data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /skills/unlock:
    post:
      operationId: unlockSkill
      tags: [skills]
      summary: Unlock a skill using skill points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockSkillBody"
      responses:
        "200":
          description: Skill unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockSkillResult"
        "400":
          description: Not enough skill points or already unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SubmitScoreBody:
      type: object
      properties:
        mode:
          type: string
        score:
          type: integer
        xpEarned:
          type: integer
        isCorrect:
          type: boolean
        responseTimeMs:
          type: ["integer", "null"]
      required:
        - mode
        - score
        - xpEarned
        - isCorrect
    ScoreResult:
      type: object
      properties:
        totalScore:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        leveledUp:
          type: boolean
        hintPoints:
          type: integer
        hintPointsEarned:
          type: integer
        rankTier:
          type: string
        skillPoints:
          type: integer
        newSkillPoint:
          type: boolean
      required:
        - totalScore
        - xp
        - level
        - leveledUp
        - hintPoints
        - hintPointsEarned
        - rankTier
        - skillPoints
        - newSkillPoint
...
    DashboardStats:
      type: object
      properties:
        totalScore:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        xpToNextLevel:
          type: integer
        gamesPlayed:
          type: integer
        phishingScore:
          type: integer
        defenseScore:
          type: integer
        builderScore:
          type: integer
        escapeScore:
          type: integer
        rank:
          type: ["integer", "null"]
        hintPoints:
          type: integer
        rankTier:
          type: string
        accuracyRate:
          type: number
        streakDays:
          type: integer
        dailyScore:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalScore
        - xp
        - level
        - xpToNextLevel
        - gamesPlayed
        - phishingScore
        - defenseScore
        - builderScore
        - escapeScore
        - rank
        - hintPoints
        - rankTier
        - accuracyRate
        - streakDays
        - dailyScore
        - recentActivity
    ActivityItem:
      type: object
      properties:
        mode:
          type: string
        score:
          type: integer
        xpEarned:
          type: integer
        createdAt:
          type: string
      required:
        - mode
        - score
        - xpEarned
        - createdAt
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [defender, attacker]
        cost:
          type: integer
        effect:
          type: string
        icon:
          type: string
      required:
        - id
        - name
        - description
        - type
        - cost
        - effect
        - icon
    SkillsResponse:
      type: object
      properties:
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        unlockedSkills:
          type: array
          items:
            type: string
        skillPoints:
          type: integer
        hackerType:
          type: string
      required:
        - skills
        - unlockedSkills
        - skillPoints
        - hackerType
    UnlockSkillBody:
      type: object
      properties:
        skillId:
          type: string
      required:
        - skillId
    UnlockSkillResult:
      type: object
      properties:
        success:
          type: boolean
        skillPoints:
          type: integer
        unlockedSkills:
          type: array
          items:
            type: string
        message:
          type: string
      required:
        - success
        - skillPoints
        - unlockedSkills
        - message
...