# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberVerse AI - Hacker vs Defender Simulator API
 * OpenAPI spec version: 0.2.0
 */

export interface ActivityItem {
  mode: string;
  score: number;
  xpEarned: number;
  createdAt: string;
}