# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberVerse AI - Hacker vs Defender Simulator API
 * OpenAPI spec version: 0.2.0
 */

export interface ScoreResult {
  totalScore: number;
  xp: number;
  level: number;
  leveledUp: boolean;
  hintPoints: number;
  hintPointsEarned: number;
  rankTier: string;
  skillPoints: number;
  newSkillPoint: boolean;
}