# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberVerse AI - Hacker vs Defender Simulator API
 * OpenAPI spec version: 0.2.0
 */
import type { SkillType } from "./skillType";

export interface Skill {
  id: string;
  name: string;
  description: string;
  type: SkillType;
  cost: number;
  effect: string;
  icon: string;
}